Relocate and rename models for transactions and calls
