+++ models/transaction.go
@@ -1,7 +1,6 @@
 package models
 
 import (
-	"bytes"
 	"encoding/hex"
 	"fmt"
 	"math/big"
@@ -10,11 +9,115 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	gethTypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/onflow/cadence"
 	cdcCommon "github.com/onflow/cadence/runtime/common"
 	"github.com/onflow/flow-go/fvm/evm/types"
 )
+
+type Transaction interface {
+	Hash() (common.Hash, error)
+	RawSignatureValues() (v *big.Int, r *big.Int, s *big.Int)
+	From() (common.Address, error)
+	To() *common.Address
+	Data() []byte
+	Nonce() uint64
+	Value() *big.Int
+	Type() uint8
+	Gas() uint64
+	GasPrice() *big.Int
+	BlobGas() uint64
+	Size() uint64
+	MarshalBinary() ([]byte, error)
+}
+
+var _ Transaction = &DirectCall{}
+
+type DirectCall struct {
+	*types.DirectCall
+}
+
+func (dc DirectCall) RawSignatureValues() (
+	v *big.Int,
+	r *big.Int,
+	s *big.Int,
+) {
+	return big.NewInt(0), big.NewInt(0), big.NewInt(0)
+}
+
+func (dc DirectCall) Hash() (common.Hash, error) {
+	return dc.DirectCall.Hash()
+}
+
+func (dc DirectCall) From() (common.Address, error) {
+	return dc.DirectCall.From.ToCommon(), nil
+}
+
+func (dc DirectCall) To() *common.Address {
+	var to *common.Address
+	if !dc.DirectCall.EmptyToField() {
+		ct := dc.DirectCall.To.ToCommon()
+		to = &ct
+	}
+	return to
+}
+
+func (dc DirectCall) Data() []byte {
+	return dc.DirectCall.Data
+}
+
+func (dc DirectCall) Nonce() uint64 {
+	return dc.DirectCall.Nonce
+}
+
+func (dc DirectCall) Value() *big.Int {
+	return dc.DirectCall.Value
+}
+
+func (dc DirectCall) Type() uint8 {
+	return dc.DirectCall.Type
+}
+
+func (dc DirectCall) Gas() uint64 {
+	return dc.DirectCall.GasLimit
+}
+
+func (dc DirectCall) GasPrice() *big.Int {
+	return dc.DirectCall.Transaction().GasPrice()
+}
+
+func (dc DirectCall) BlobGas() uint64 {
+	return dc.DirectCall.Transaction().BlobGas()
+}
+
+func (dc DirectCall) Size() uint64 {
+	return dc.DirectCall.Transaction().Size()
+}
+
+func (dc DirectCall) MarshalBinary() ([]byte, error) {
+	return dc.DirectCall.Encode()
+}
+
+var _ Transaction = &TransactionCall{}
+
+type TransactionCall struct {
+	*gethTypes.Transaction
+}
+
+func (tc TransactionCall) Hash() (common.Hash, error) {
+	return tc.Transaction.Hash(), nil
+}
+
+func (tc TransactionCall) From() (common.Address, error) {
+	return gethTypes.Sender(
+		gethTypes.LatestSignerForChainID(tc.ChainId()),
+		tc.Transaction,
+	)
+}
+
+func (tc TransactionCall) MarshalBinary() ([]byte, error) {
+	encoded, err := tc.Transaction.MarshalBinary()
+	return append([]byte{tc.Type()}, encoded...), err
+}
 
 type txEventPayload struct {
 	BlockHeight             uint64 `cadence:"blockHeight"`
@@ -34,66 +137,11 @@ func (tx *txEventPayload) IsDirectCall() bool {
 	return tx.TransactionType == types.DirectCallTxType
 }
 
-// DecodeReceipt takes a cadence event for transaction executed and decodes it into the receipt.
-func DecodeReceipt(event cadence.Event) (*gethTypes.Receipt, error) {
-	if !IsTransactionExecutedEvent(event) {
-		return nil, fmt.Errorf(
-			"invalid event type for decoding into receipt, received %s expected %s",
-			event.Type().ID(),
-			types.EventTypeTransactionExecuted,
-		)
-	}
-
-	var tx txEventPayload
-	err := cadence.DecodeFields(event, &tx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to cadence decode receipt: %w", err)
-	}
-
-	encLogs, err := hex.DecodeString(tx.Logs)
-	if err != nil {
-		return nil, fmt.Errorf("failed to hex decode receipt: %w", err)
-	}
-
-	var logs []*gethTypes.Log
-	if len(encLogs) > 0 {
-		err = rlp.Decode(bytes.NewReader(encLogs), &logs)
-		if err != nil {
-			return nil, fmt.Errorf("failed to rlp decode receipt: %w", err)
-		}
-	}
-
-	receipt := &gethTypes.Receipt{
-		BlockNumber:       big.NewInt(int64(tx.BlockHeight)),
-		Type:              uint8(tx.TransactionType),
-		Logs:              logs,
-		TxHash:            common.HexToHash(tx.TransactionHash),
-		ContractAddress:   common.HexToAddress(tx.DeployedContractAddress),
-		GasUsed:           tx.GasConsumed,
-		CumulativeGasUsed: tx.GasConsumed, // todo check
-		EffectiveGasPrice: nil,            // todo check
-		BlobGasUsed:       0,              // todo check
-		BlobGasPrice:      nil,            // todo check
-		TransactionIndex:  0,              // todo add tx index in evm core event
-		BlockHash:         common.HexToHash(tx.BlockHash),
-	}
-
-	if tx.Failed {
-		receipt.Status = gethTypes.ReceiptStatusFailed
-	} else {
-		receipt.Status = gethTypes.ReceiptStatusSuccessful
-	}
-
-	receipt.Bloom = gethTypes.CreateBloom([]*gethTypes.Receipt{receipt})
-
-	return receipt, nil
-}
-
 // DecodeTransaction takes a cadence event for transaction executed
 // and decodes it into a geth transaction. If the transaction was
 // generated by a direct call, we also return it.
 func DecodeTransaction(event cadence.Event) (
-	FlowEVMTxData,
+	Transaction,
 	error,
 ) {
 	if !IsTransactionExecutedEvent(event) {
@@ -115,7 +163,7 @@ func DecodeTransaction(event cadence.Event) (
 		return nil, fmt.Errorf("failed to decode transaction hex: %w", err)
 	}
 
-	var evmTxData FlowEVMTxData
+	var tx Transaction
 
 	// check if the transaction data is actually from a direct call,
 	// which is a special state transition in Flow EVM.
@@ -125,17 +173,17 @@ func DecodeTransaction(event cadence.Event) (
 			return nil, fmt.Errorf("failed to rlp decode direct call: %w", err)
 		}
 
-		evmTxData = DirectCallTx{DirectCall: directCall}
+		tx = DirectCall{DirectCall: directCall}
 
-		return evmTxData, nil
+		return tx, nil
 	} else {
-		tx := &gethTypes.Transaction{}
-		if err := tx.UnmarshalBinary(encodedTx); err != nil {
+		gethTx := &gethTypes.Transaction{}
+		if err := gethTx.UnmarshalBinary(encodedTx); err != nil {
 			return nil, fmt.Errorf("failed to rlp decode transaction: %w", err)
 		}
-		evmTxData = GethTx{Transaction: tx}
+		tx = TransactionCall{Transaction: gethTx}
 
-		return evmTxData, nil
+		return tx, nil
 	}
 }
